Refactor: 도메인 관련 클래스 수정
